+++ database/seeds/AirfieldTableSeeder.php
@@ -1,7 +1,20 @@
 <?php
 
+use App\Models\Airfield;
+use Carbon\Carbon;
+use Illuminate\Database\Seeder;
 
-class AirfieldTableSeeder
+class AirfieldTableSeeder extends Seeder
 {
-
+    public function run()
+    {
+        Airfield::create(
+            [
+                'code' => 'EGLL',
+                'transition_altitude' => 6000,
+                'created_at' => Carbon::now()->subHour(),
+                'updated_at' => Carbon::now(),
+            ]
+        );
+    }
 }
