+++ database/seeds/TmaMslCalculationTableSeeder.php
@@ -1,16 +1,16 @@
 <?php
 
-use App\Models\AirfieldMslCalculation;
+use App\Models\TmaMslCalculation;
 use Illuminate\Database\Seeder;
 
-class AirfieldMslCalculationTableSeeder extends Seeder
+class TmaMslCalculationTableSeeder extends Seeder
 {
     public function run()
     {
-        AirfieldMslCalculation::create(
+        TmaMslCalculation::create(
             [
-                'airfield_id' => 1,
-                'calculation' => ['type' => 'airfield', 'code' => 'EGLL'],
+                'tma_id' => 1,
+                'calculation' => json_encode(['type' => 'airfield', 'code' => 'EGLL']),
             ]
         );
     }
